[spark] Fix self-referencing RTAS data loss and extract PaimonTableAsSelectHelper
- Pin snapshot in query for self-referencing RTAS (CREATE OR REPLACE
  TABLE t AS SELECT * FROM t) to prevent reading truncated data.
  Relations with user-specified time travel options are left unchanged.
- Extract PaimonTableAsSelectHelper from PaimonReplaceTableStrategyHelper
  to share isV1SaveAsTableOverwrite, pinSnapshotInQuery, rewriteToOverwrite
  across Spark versions without duplication.
- Fix 'save as partitioned table' test: guard primaryKeys assertion with
  gteqSpark3_4 since Spark 3.2/3.3 loses PK on replace.
