+++ paimon-spark/paimon-spark-3.2/src/main/scala/org/apache/spark/sql/execution/shim/PaimonCreateTableAsSelectStrategy.scala
@@ -24,7 +24,8 @@ import org.apache.paimon.spark.catalog.FormatTableCatalog
 import org.apache.spark.sql.{SparkSession, Strategy}
 import org.apache.spark.sql.catalyst.plans.logical.{CreateTableAsSelect, LogicalPlan}
 import org.apache.spark.sql.connector.catalog.CatalogV2Util
-import org.apache.spark.sql.execution.{PaimonStrategyHelper, SparkPlan}
+import org.apache.spark.sql.execution.{PaimonTableAsSelectHelper, SparkPlan}
+import org.apache.spark.sql.execution.PaimonTableAsSelectHelper._
 import org.apache.spark.sql.execution.datasources.v2.CreateTableAsSelectExec
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
 
@@ -41,7 +42,8 @@ case class PaimonCreateTableAsSelectStrategy(spark: SparkSession) extends Strate
           props,
           options,
           ifNotExists) =>
-      val (tableOptions, writeOptions) = PaimonStrategyHelper.splitTableAndWriteOptions(options)
+      val (tableOptions, writeOptions) =
+        splitTableAndWriteOptions(options)
       val newProps = CatalogV2Util.withDefaultOwnership(props) ++ tableOptions
 
       val isPartitionedFormatTable = {
